Created toolbar for HomeView, and gave ActionButton an animation.
