--- src/SmartJournal/SmartJournal/Components/CustomButton/CustomButton.swift
@@ -1,20 +1,0 @@
-//
-//  CustomButton.swift
-//  SmartJournal
-//
-//  Created by Usman Tahir Qureshi on 11/9/23.
-//
-
-import SwiftUI
-
-struct ActionButton: ButtonStyle {
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .frame(maxWidth: .infinity)
-            .padding()
-            .background(Color.actionButton)
-            .foregroundColor(.white)
-            .cornerRadius(10)
-        
-    }
-}
